Fix issue with the starting address of an allocation request
Took a while to figure out why a CPU address exception was being
thrown when the address itself is clearly valid (R/W). The issue was
because the header is 2 bytes, the address returned to the requestor
will always be on a 2-byte boundary. This causes alignment issues if a
long-write is performed. The solution is to pad the header to 4-bytes.

+++ libyaul/kernel/mm/slob.c
@@ -23,7 +23,10 @@ STATIC_ASSERT(SLOB_PAGE_SIZE >= SLOB_PAGE_BREAK_2ND);
 
 struct slob_block {
         union {
-                int16_t sb_bunits;
+                /* The start of the address of the request returned to
+                 * the requestor must be on a 4 byte boundary due to
+                 * alignment issues. */
+                int32_t sb_bunits;
                 /* Next "member" is a link to the next block relative to
                  * the starting address of the SLOB page. */
                 int16_t sb_bnext;
